+++ meta-processor/src/main/kotlin/net/skyscanner/backpack/ksp/References.kt
@@ -32,7 +32,6 @@ object ViewStoryAnnotation : AnnotationDefinition {
   override val pkg = metaPkg
   val paramName = stringParamOf("name")
   val paramScreenshot = booleanParamOf("screenshot")
-  val paramLayoutId = intParamOf("layoutId")
 }
 
 object ComponentAnnotation : AnnotationDefinition {
